requirements.txt
streamlit
requests
Pillow
PyMuPDF